+++ Project_2_Web_Api/Project_2_Web_Api/Controllers/PositionGroupController.cs
@@ -9,13 +9,16 @@ using Project_2_Web_API.Models;
 namespace Project_2_Web_Api.Controllers;
 [Route("api/[controller]")]
 [ApiController]
+[Authorize]
 public class PositionGroupController : ControllerBase
 {
 
 	private readonly PositionGroupService positionGroupService;
-	public PositionGroupController(PositionGroupService positionGroupService)
+	private readonly UserServiceAccessor userServiceAccessor;
+	public PositionGroupController(PositionGroupService positionGroupService, UserServiceAccessor userServiceAccessor)
 	{
 		this.positionGroupService = positionGroupService;
+		this.userServiceAccessor = userServiceAccessor;
 	}
 
 
@@ -26,6 +29,18 @@ public class PositionGroupController : ControllerBase
 	{
 		try
 		{
+			if (await userServiceAccessor.IsGuest())
+			{
+				return Unauthorized();
+			}
+			if (await userServiceAccessor.IsDistributor())
+			{
+				return Unauthorized();
+			}
+			if (await userServiceAccessor.IsSales())
+			{
+				return Unauthorized();
+			}
 			return  Ok( await positionGroupService.FindAll());
 		}
 		catch(Exception ex)
@@ -43,6 +58,18 @@ public class PositionGroupController : ControllerBase
 	{
 		try
 		{
+			if (await userServiceAccessor.IsGuest())
+			{
+				return Unauthorized();
+			}
+			if (await userServiceAccessor.IsDistributor())
+			{
+				return Unauthorized();
+			}
+			if (await userServiceAccessor.IsSales())
+			{
+				return Unauthorized();
+			}
 			return Ok(await positionGroupService.FindById(id));
 		}
 		catch (Exception ex)
@@ -58,6 +85,18 @@ public class PositionGroupController : ControllerBase
 	{
 		try
 		{
+			if (await userServiceAccessor.IsGuest())
+			{
+				return Unauthorized();
+			}
+			if (await userServiceAccessor.IsDistributor())
+			{
+				return Unauthorized();
+			}
+			if (await userServiceAccessor.IsSales())
+			{
+				return Unauthorized();
+			}
 			return Ok(await positionGroupService.FindByName(name));
 		}
 		catch (Exception ex)
@@ -72,6 +111,18 @@ public class PositionGroupController : ControllerBase
 	[HttpPost("create")]
 	public async Task<IActionResult> Create([FromBody]PositionGroupDTO request)
 	{
+		if (await userServiceAccessor.IsGuest())
+		{
+			return Unauthorized();
+		}
+		if (await userServiceAccessor.IsDistributor())
+		{
+			return Unauthorized();
+		}
+		if (await userServiceAccessor.IsSales())
+		{
+			return Unauthorized();
+		}
 		return await positionGroupService.Create(request);
 	}
 	
@@ -81,6 +132,18 @@ public class PositionGroupController : ControllerBase
 	[HttpPut("update/{id}")]
 	public async Task<IActionResult> Update(string id, [FromBody] PositionGroupDTO request)
 	{
+		if (await userServiceAccessor.IsGuest())
+		{
+			return Unauthorized();
+		}
+		if (await userServiceAccessor.IsDistributor())
+		{
+			return Unauthorized();
+		}
+		if (await userServiceAccessor.IsSales())
+		{
+			return Unauthorized();
+		}
 		return await positionGroupService.Update(id, request);
 	}
 
@@ -91,6 +154,18 @@ public class PositionGroupController : ControllerBase
 	[HttpDelete("delete/{id}")]
 	public async Task<IActionResult> Delete(string id)
 	{
+		if (await userServiceAccessor.IsGuest())
+		{
+			return Unauthorized();
+		}
+		if (await userServiceAccessor.IsDistributor())
+		{
+			return Unauthorized();
+		}
+		if (await userServiceAccessor.IsSales())
+		{
+			return Unauthorized();
+		}
 		return await positionGroupService.Delete(id);
 	}
 }
