+++ Project_2_Web_Api/Project_2_Web_Api/Controllers/UserController.cs
@@ -1,18 +1,23 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using Project_2_Web_Api.DTO;
 using Project_2_Web_Api.Service;
+using System.Security.Claims;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
 namespace Project_2_Web_Api.Controllers;
 [Route("api/[controller]")]
 [ApiController]
+[Authorize]
 public class UserController : ControllerBase
 {
 	private readonly UserService userService;
-	public UserController(UserService userService)
+	private readonly UserServiceAccessor userServiceAccessor;
+	public UserController(UserService userService, UserServiceAccessor userServiceAccessor)
 	{
 		this.userService = userService;
+		this.userServiceAccessor = userServiceAccessor;
 	}
 	[Produces("application/json")]
 	[Consumes("application/json")]
@@ -21,6 +26,17 @@ public class UserController : ControllerBase
 	{
 		try
 		{
+			if(await userServiceAccessor.IsGuest())
+			{
+				return Unauthorized();
+			}if(await userServiceAccessor.IsDistributor())
+			{
+				return Unauthorized();
+			}
+			if (await userServiceAccessor.IsSales())
+			{
+				return Unauthorized();
+			}
 			return Ok(await userService.FindAll());
 		}catch(Exception ex)
 		{
@@ -35,6 +51,18 @@ public class UserController : ControllerBase
 	{
 		try
 		{
+			if (await userServiceAccessor.IsGuest())
+			{
+				return Unauthorized();
+			}
+			if (await userServiceAccessor.IsDistributor())
+			{
+				return Unauthorized();
+			}
+			if (await userServiceAccessor.IsSales())
+			{
+				return Unauthorized();
+			}
 			return Ok(await userService.FindById(id));
 		}
 		catch (Exception ex)
@@ -50,6 +78,18 @@ public class UserController : ControllerBase
 	{
 		try
 		{
+			if (await userServiceAccessor.IsGuest())
+			{
+				return Unauthorized();
+			}
+			if (await userServiceAccessor.IsDistributor())
+			{
+				return Unauthorized();
+			}
+			if (await userServiceAccessor.IsSales())
+			{
+				return Unauthorized();
+			}
 			return Ok(await userService.FindByName(name));
 		}
 		catch (Exception ex)
@@ -63,6 +103,18 @@ public class UserController : ControllerBase
 	[HttpPost("create")]
 	public async Task<IActionResult> Create([FromBody] UserDTO request)
 	{
+		if (await userServiceAccessor.IsGuest())
+		{
+			return Unauthorized();
+		}
+		if (await userServiceAccessor.IsDistributor())
+		{
+			return Unauthorized();
+		}
+		if (await userServiceAccessor.IsSales())
+		{
+			return Unauthorized();
+		}
 		return await userService.Create(request);
 	}
 
@@ -71,6 +123,18 @@ public class UserController : ControllerBase
 	[HttpPut("update/{id}")]
 	public async Task<IActionResult> Update(string id, [FromBody] UserDTO request)
 	{
+		if (await userServiceAccessor.IsGuest())
+		{
+			return Unauthorized();
+		}
+		if (await userServiceAccessor.IsDistributor())
+		{
+			return Unauthorized();
+		}
+		if (await userServiceAccessor.IsSales())
+		{
+			return Unauthorized();
+		}
 		return await userService.Update(id, request);
 	}
 
@@ -79,6 +143,18 @@ public class UserController : ControllerBase
 	[HttpDelete("delete/{id}")]
 	public async Task<IActionResult> Delete(string id)
 	{
+		if (await userServiceAccessor.IsGuest())
+		{
+			return Unauthorized();
+		}
+		if (await userServiceAccessor.IsDistributor())
+		{
+			return Unauthorized();
+		}
+		if (await userServiceAccessor.IsSales())
+		{
+			return Unauthorized();
+		}
 		return await userService.Delete(id);
 	}
 	[Produces("application/json")]
@@ -88,6 +164,18 @@ public class UserController : ControllerBase
 	{
 		try
 		{
+			if (await userServiceAccessor.IsGuest())
+			{
+				return Unauthorized();
+			}
+			if (await userServiceAccessor.IsDistributor())
+			{
+				return Unauthorized();
+			}
+			if (await userServiceAccessor.IsSales())
+			{
+				return Unauthorized();
+			}
 			return await userService.SettingPermission(request.Id, request.PermissionId);
 		}
 		catch (Exception ex)
@@ -101,6 +189,45 @@ public class UserController : ControllerBase
 	[HttpPut("reset-password")]
 	public async Task<IActionResult> ResetPassword([FromBody] ResetPasswordRequest request)
 	{
+		if (await userServiceAccessor.IsGuest())
+		{
+			return Unauthorized();
+		}
+		if (await userServiceAccessor.IsDistributor())
+		{
+			return Unauthorized();
+		}
+		if (await userServiceAccessor.IsSales())
+		{
+			return Unauthorized();
+		}
 		return await userService.ResetPassword(request.Id);
+	}
+
+	[Produces("application/json")]
+	[Consumes("application/json")]
+	[HttpGet("get-me")]
+	public async Task<IActionResult> GetMe()
+	{
+		try
+		{
+			if (await userServiceAccessor.IsDistributor())
+			{
+				return Unauthorized();
+			}
+			if (await userServiceAccessor.IsSales())
+			{
+				return Unauthorized();
+			}
+			if(await userServiceAccessor.IsSystem())
+			{
+				return Unauthorized();
+			}
+			return Ok(await userService.FindById(User?.Identity?.Name));
+		}
+		catch (Exception ex)
+		{
+			return BadRequest(new { error = ex.Message });
+		}
 	}
 }
